Move IsCompileTimeConstantPass to separate file
